Add permission check for critical endpoints

+++ src/main/java/com/example/tcc/controllers/AssetController.java
@@ -24,9 +24,14 @@ public class AssetController {
     private final AssetDeletionService assetDeletionService;
     private final AssetImageAdditionService assetImageAdditionService;
     private final AssetImageDeletionService assetImageDeletionService;
+    private final PermissionCheckService permissionCheckService;
 
     @PostMapping
     public ResponseEntity<CreateAssetResponseDto> create(@RequestParam Long fileId, @RequestParam MultipartFile image) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if(!permissionCheckService.checkPermissionForFile((Long)authentication.getPrincipal(), fileId))
+            return ResponseEntity.status(401).build();
+
         CreateAssetResponseDto response = assetCreationService.createAndRecognize(fileId, image);
         return ResponseEntity.ok(response);
     }
@@ -34,12 +39,19 @@ public class AssetController {
     @PostMapping("/confirm/{assetId}")
     public ResponseEntity<Void> confirm(@PathVariable Long assetId, @RequestBody AssetConfirmationRequestDto requestDto) {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if(!permissionCheckService.checkPermissionForAsset((Long)authentication.getPrincipal(), assetId))
+            return ResponseEntity.status(401).build();
+
         assetConfirmationService.confirm(authentication.getCredentials().toString(), requestDto.getAssetNumber(), assetId);
         return ResponseEntity.ok().build();
     }
 
     @PostMapping("/add-image")
     public ResponseEntity<Map<String, String>> add(@RequestParam Long assetId, @RequestParam MultipartFile image) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if(!permissionCheckService.checkPermissionForAsset((Long)authentication.getPrincipal(), assetId))
+            return ResponseEntity.status(401).build();
+
         String filepath = assetImageAdditionService.add(assetId, image);
         Map<String, String> response = new HashMap<>();
         response.put("path", filepath);
@@ -48,12 +60,20 @@ public class AssetController {
 
     @DeleteMapping("/delete-image/{filename}")
     public ResponseEntity<Void> deleteImage(@PathVariable String filename) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if(!permissionCheckService.checkPermissionForImageFilename((Long)authentication.getPrincipal(), filename))
+            return ResponseEntity.status(401).build();
+
         assetImageDeletionService.delete(filename);
         return ResponseEntity.noContent().build();
     }
 
     @DeleteMapping("/{assetId}")
     public ResponseEntity<Void> deleteAsset(@PathVariable Long assetId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if(!permissionCheckService.checkPermissionForAsset((Long)authentication.getPrincipal(), assetId))
+            return ResponseEntity.status(401).build();
+
         assetDeletionService.delete(assetId);
         return ResponseEntity.noContent().build();
     }
